Move SuiteInfoRecord class to suite_runner; Add a VERSION attribute to BaseSuite and record it to test summary
