+++ tests/mobly/records_test.py
@@ -542,21 +542,6 @@ class RecordsTest(unittest.TestCase):
 
     self.assertEqual(test_uid_helper.uid, 'some-uuid')
 
-  def test_convert_suite_info_record_to_dict(self):
-    suite_class_name = 'FakeTestSuite'
-    record = records.SuiteInfoRecord(test_suite_class=suite_class_name)
-    record.suite_begin()
-    record.suite_end()
-
-    result = record.to_dict()
-
-    self.assertIn(
-        (records.SuiteInfoRecord.KEY_TEST_SUITE_CLASS, suite_class_name),
-        result.items(),
-    )
-    self.assertIn(records.SuiteInfoRecord.KEY_BEGIN_TIME, result)
-    self.assertIn(records.SuiteInfoRecord.KEY_END_TIME, result)
-
 
 if __name__ == '__main__':
   unittest.main()
